Drop unused code and deps
Nothing uses ZMQ_CONTEXT anymore, which was also the only user of
lazy_static package. Drop them.

Also move some code to the only place it is used.
