+++ components/butterfly/src/lib.rs
@@ -46,8 +46,6 @@
 extern crate byteorder;
 extern crate habitat_core;
 #[macro_use]
-extern crate lazy_static;
-#[macro_use]
 extern crate log;
 extern crate protobuf;
 extern crate rand;
@@ -69,30 +67,4 @@ pub mod message;
 pub mod rumor;
 pub mod server;
 
-use std::cell::UnsafeCell;
-
 pub use server::Server;
-
-lazy_static! {
-    /// A threadsafe shared ZMQ context for consuming services.
-    ///
-    /// You probably want to use this context to create new ZMQ sockets unless you *do not* want to
-    /// connect them together using an in-proc queue.
-    pub static ref ZMQ_CONTEXT: Box<ServerContext> = {
-        let ctx = ServerContext(UnsafeCell::new(zmq::Context::new()));
-        Box::new(ctx)
-    };
-}
-
-/// This is a wrapper to provide interior mutability of an underlying `zmq::Context` and allows
-/// for sharing/sending of a `zmq::Context` between threads.
-pub struct ServerContext(UnsafeCell<zmq::Context>);
-
-impl ServerContext {
-    pub fn as_mut(&self) -> &mut zmq::Context {
-        unsafe { &mut *self.0.get() }
-    }
-}
-
-unsafe impl Send for ServerContext {}
-unsafe impl Sync for ServerContext {}
